updated rpsec rails to new helper filenames
