+++ Gemfile
@@ -79,6 +79,7 @@ group :test do
   gem 'factory_girl_rails'
   gem 'coveralls', require: false
   gem 'poltergeist'
+  gem 'capybara-screenshot'
 end
 
 group :development do
