+++ spec/controllers/comments_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe CommentsController, type: :controller do
   let!(:user) { build(:user) }
