+++ spec/controllers/password_resets_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe PasswordResetsController, type: :controller do
