+++ spec/features/apply_to_adopt_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature 'Apply for Adoption' do
   scenario "Renter fills out adoption application", js: true do
