+++ spec/features/home_page_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 feature "Home Page" do
   scenario "visitor can view events" do
