+++ spec/models/adopter_searcher_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe AdopterSearcher do
   describe '.search' do
