+++ spec/models/adoption_spec.rb
@@ -10,7 +10,7 @@
 #  updated_at    :timestamp(6)
 #
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe Adoption do
   let(:adoption) { build(:adoption) }
