+++ spec/models/dog_searcher_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe DogSearcher do
   describe '.search' do
