+++ spec/models/user_spec.rb
@@ -52,7 +52,7 @@
 #  mentor_id              :integer
 #
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe User do
   before do
